Adding Compose File for Emby
I built the image, tested it with my server and it worked. I did not see any errors on the worker/flask-app.
If you can go through the code, and test it with the other types of servers to make sure nothing is broken it would be very helpful. As I said I have very little experience with python and I only compared  the implimented Jellyfin API calls to EMBY.SDK Python library/API Documentations.