working Open AI Integration
To test this make sure Docker is up to date and that docker-model-plugin is installed.

I added the correct configurations to access LLM Model from inside the container. It is OpenAI API compatible. I used OpenAI SDK Python Library to impliment the function.

The LLM included is about 500mb and on a labtop processor i5 it took around 20s to return a result.